# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TurnByTurn API",
    "version": "1.0.0",
    "description": "Backend for TurnByTurn, a rotating savings (ajo / esusu) app. All money values are integer kobo. Errors always return `{ error: { code, message, details?, retryAfter? } }`; see docs/ERRORS.md for the code → screen map."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "CONFLICT",
                  "INTERNAL_ERROR",
                  "GATEWAY_ERROR",
                  "INVALID_CREDENTIALS",
                  "OTP_INVALID",
                  "OTP_EXPIRED",
                  "OTP_RESEND_COOLDOWN",
                  "PASSWORD_POLICY",
                  "ACCOUNT_EXISTS",
                  "TOO_MANY_ATTEMPTS",
                  "BANK_VERIFICATION_FAILED",
                  "SESSION_EXPIRED",
                  "UNAUTHENTICATED",
                  "ACCESS_DENIED",
                  "SIGNUP_INCOMPLETE",
                  "GROUP_NOT_DRAFT",
                  "GROUP_NOT_ACTIVE",
                  "GROUP_SETUP_INCOMPLETE",
                  "GROUP_FULL",
                  "INVITE_INVALID",
                  "ALREADY_MEMBER",
                  "AMOUNT_INVALID",
                  "NOTHING_DUE",
                  "CYCLE_CLOSED",
                  "PAYOUT_NOT_ELIGIBLE",
                  "PAYOUT_ACCOUNT_MISSING",
                  "REMINDER_COOLDOWN"
                ]
              },
              "message": {
                "type": "string"
              },
              "retryAfter": {
                "type": "integer",
                "description": "Seconds; also sent as Retry-After header"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  },
  "paths": {
...
    "/banks": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Nigerian banks (Squadco bank codes)",
        "operationId": "getBanks",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "banks": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/content/help": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Payment Help FAQ + support contacts",
        "operationId": "getContentHelp",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/invites/{code}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Public invite teaser (marketing site / deep link)",
        "operationId": "getInvitesCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "INVITE_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/files/{id}": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Profile photo bytes",
        "operationId": "getFilesId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}